requirements.txt
bson==0.5.10
click==8.0.4
colorama==0.4.5
dataclasses==0.6
dnspython==2.2.1
Flask==2.0.3
Flask-Cors==4.0.0
gunicorn==21.2.0
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
pymongo==4.1.1
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.20.0
six==1.16.0
typing-extensions==4.1.1
Werkzeug==2.0.3
zipp==3.6.0
